Popup menu splash doesn't go to edge
Now the splash goes to the edge. Also, fold popup_menu_item.dart into
popup_menu.dart for simplicity.

Fixes #266
