--- packages/flutter/lib/src/material/popup_menu_item.dart
@@ -1,34 +1,0 @@
-// Copyright 2015 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'package:flutter/widgets.dart';
-
-import 'theme.dart';
-
-const double _kMenuItemHeight = 48.0;
-const double _kBaselineOffsetFromBottom = 20.0;
-
-class PopupMenuItem<T> extends StatelessComponent {
-  PopupMenuItem({
-    Key key,
-    this.value,
-    this.child
-  }) : super(key: key);
-
-  final Widget child;
-  final T value;
-
-  Widget build(BuildContext context) {
-    return new Container(
-      height: _kMenuItemHeight,
-      child: new DefaultTextStyle(
-        style: Theme.of(context).text.subhead,
-        child: new Baseline(
-          baseline: _kMenuItemHeight - _kBaselineOffsetFromBottom,
-          child: child
-        )
-      )
-    );
-  }
-}
